+++ requests/jobs.http
@@ -2,7 +2,7 @@ GET http://localhost:5000/api/jobs?skip=6&limit=3
 
 ###
 
-GET http://localhost:5000/api/jobs/62c9f2b778a9ce77381dd56a
+GET http://localhost:5000/api/jobs/62d88ef33721c4e5359fd1c2
 
 ###
 
@@ -12,7 +12,7 @@ GET http://localhost:5000/api/jobs/search?title=&skip=0&limit=5
 
 POST http://localhost:5000/api/jobs
 Content-Type: application/json
-Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiI2MmNjODVlN2M0YWMxNTEwNDdhOGQzM2YiLCJpYXQiOjE2NTc1NzQzMTIsImV4cCI6MTY1NzU3NzkxMn0.3d5lzLZCwXVnEUxVQneEDrD324qoIokJ005Wm9f15jI
+Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiI2MmNjODVlN2M0YWMxNTEwNDdhOGQzM2YiLCJpYXQiOjE2NTgzNTk1MTMsImV4cCI6MTY1ODM2MzExM30.cxvIm61DX-iIzlNToJuHWKLSgAYkKUgbGV-EtSC3h4M
 
 { 
     "title": "title to update"
@@ -25,9 +25,9 @@ Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiI2MmNjO
 
 ###
 
-PATCH http://localhost:5000/api/jobs/62cc959fd6012552a167e85b
+PATCH http://localhost:5000/api/jobs/62d88ef33721c4e5359fd1c2
 Content-Type: application/json
-Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiI2MmNjODVlN2M0YWMxNTEwNDdhOGQzM2YiLCJpYXQiOjE2NTc1NzQzMTIsImV4cCI6MTY1NzU3NzkxMn0.3d5lzLZCwXVnEUxVQneEDrD324qoIokJ005Wm9f15jI
+Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiI2MmNjODVlN2M0YWMxNTEwNDdhOGQzM2YiLCJpYXQiOjE2NTgzNTk1MTMsImV4cCI6MTY1ODM2MzExM30.cxvIm61DX-iIzlNToJuHWKLSgAYkKUgbGV-EtSC3h4M
 
 { 
     "title": "title update"
